Move Actionlike macro into own file
